Add more comments about locking, correct typo, move BitHashSet to util folder
