+++ src/java/main/org/apache/zookeeper/server/watch/IDeadWatcherListener.java
@@ -23,7 +23,7 @@ import java.util.Set;
 /**
  * Interface used to process the dead watchers related to closed cnxns.
  */
-public interface DeadWatcherListener {
+public interface IDeadWatcherListener {
 
     /**
      * Process the given dead watchers.
