+++ src/java/test/org/apache/zookeeper/server/util/BitHashSetTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.zookeeper.server.watch;
+package org.apache.zookeeper.server.util;
 
 import java.util.Set;
 import java.util.HashSet;
